Renamed KitWalletProvider to SequenceWalletProvider
