+++ examples/react/src/App.tsx
@@ -1,7 +1,7 @@
 import { ThemeProvider } from '@0xsequence/design-system'
 import { SequenceCheckoutProvider } from '@0xsequence/web-sdk-checkout'
 import { SequenceKit } from '@0xsequence/web-sdk-connect'
-import { KitWalletProvider } from '@0xsequence/web-sdk-wallet'
+import { SequenceWalletProvider } from '@0xsequence/web-sdk-wallet'
 
 import { Homepage } from './components/Homepage'
 import { config } from './config'
@@ -10,11 +10,11 @@ export const App = () => {
   return (
     <ThemeProvider theme="dark">
       <SequenceKit config={config}>
-        <KitWalletProvider>
+        <SequenceWalletProvider>
           <SequenceCheckoutProvider>
             <Homepage />
           </SequenceCheckoutProvider>
-        </KitWalletProvider>
+        </SequenceWalletProvider>
       </SequenceKit>
     </ThemeProvider>
   )
