+++ packages/wallet/src/components/CoinRow.tsx
