+++ packages/wallet/src/components/CollectibleTileImage.tsx
