+++ packages/wallet/src/components/CopyButton.tsx
