+++ packages/wallet/src/components/DefaultIcon.tsx
