+++ packages/wallet/src/components/NetworkBadge.tsx
