+++ packages/wallet/src/components/SequenceWalletProvider/index.ts
@@ -1,0 +1,1 @@
+export * from './SequenceWalletProvider'
