+++ packages/wallet/src/index.ts
@@ -1,5 +1,5 @@
 // Provider
-export { KitWalletProvider } from './shared/KitWalletProvider'
+export { SequenceWalletProvider } from './components/SequenceWalletProvider'
 
 // Hooks
 export { useOpenWalletModal } from './hooks/useOpenWalletModal'
