--- packages/wallet/src/shared/index.ts
@@ -1,4 +1,0 @@
-export * from './KitWalletProvider'
-export { CollectibleTileImage } from './CollectibleTileImage'
-export { useSettings } from '../hooks'
-export { NetworkBadge } from './NetworkBadge'
