+++ packages/wallet/src/views/History.tsx
@@ -3,7 +3,7 @@ import React from 'react'
 import { useAccount } from 'wagmi'
 
 import { useSettings } from '../hooks'
-import { TransactionHistoryList } from '../shared/TransactionHistoryList'
+import { TransactionHistoryList } from '../components/TransactionHistoryList'
 
 export const History = () => {
   const { selectedNetworks } = useSettings()
