+++ packages/wallet/src/views/Home/components/AssetSummary/CollectibleTile/index.tsx
@@ -1,7 +1,7 @@
 import { TokenBalance } from '@0xsequence/indexer'
 import { useGetTokenMetadata } from '@0xsequence/web-sdk-hooks'
 
-import { CollectibleTileImage } from '../../../../../shared/CollectibleTileImage'
+import { CollectibleTileImage } from '../../../../../components/CollectibleTileImage'
 
 interface CollectibleTileProps {
   balance: TokenBalance
